requirements.txt
Django==5.0
neomodel==5.1.2
django-neomodel==0.1.1
djangorestframework==3.14.0